Modify layout

+++ BigInteger/18127087_18127207/QInt.cpp
@@ -402,14 +402,16 @@ QInt QInt::operator+(const QInt& Qi)
 
 // Unary Operators
 // Support Operators
-QInt& QInt::operator=(const QInt& Qi) {
+QInt& QInt::operator=(const QInt& Qi) 
+{
 	QInt tempQi;
 	tempQi.arrayBits = Qi.arrayBits;
 	return tempQi;
 }
 
 // Bitwise
-QInt QInt::operator~() {
+QInt QInt::operator~() 
+{
 	QInt tempQi;
 	tempQi.arrayBits.flip();
 	return tempQi;
@@ -417,38 +419,49 @@ QInt QInt::operator~() {
 
 // Binary Operators
 // Bitwise
-QInt QInt::operator<<(const int8_t&) {
-	for (int i = this->arrayBits.size() - 1; i >= 0; i--) {
+QInt QInt::operator<<(const int8_t&) 
+{
+	for (int i = this->arrayBits.size() - 1; i >= 0; i--) 
+	{
 		this->arrayBits[i] << 1;
 	}
 	return *this;
 }
 
-QInt QInt::operator>>(const int8_t&) {
-	for (int i = this->arrayBits.size() - 2; i >= 0; i--) {
+QInt QInt::operator>>(const int8_t&) 
+{
+	for (int i = this->arrayBits.size() - 2; i >= 0; i--) 
+	{
 		this->arrayBits[i] >> 1;
 	}
 	return *this;
 }
-QInt QInt::operator&(const QInt& Qi) {
+
+QInt QInt::operator&(const QInt& Qi) 
+{
 	QInt tempQi;
-	for (int i = 0; i < Qi.arrayBits.size(); i++) {
+	for (int i = 0; i < (int)Qi.arrayBits.size(); i++) 
+	{
 		tempQi.arrayBits[i] = this->arrayBits[i] & Qi.arrayBits[i];
 	}
 	return tempQi;
 }
 
-QInt QInt::operator|(const QInt& Qi) {
+QInt QInt::operator|(const QInt& Qi) 
+{
 	QInt tempQi;
-	for (int i = 0; i < Qi.arrayBits.size(); i++) {
+	for (int i = 0; i < (int)Qi.arrayBits.size(); i++) 
+	{
 		tempQi.arrayBits[i] = this->arrayBits[i] | Qi.arrayBits[i];
 	}
 	return tempQi;
 }
 
-QInt QInt::operator^(const QInt& Qi) {
+QInt QInt::operator^(const QInt& Qi) 
+{
 	QInt tempQi;
-	for (int i = 0; i < Qi.arrayBits.size(); i++) {
+	for (int i = 0; i < (int)Qi.arrayBits.size(); i++) 
+	{
 		tempQi.arrayBits[i] = this->arrayBits[i] ^ Qi.arrayBits[i];
 	}
 	return tempQi;
